change youtube search feature logic
