+++ apps/server/.gitignore
@@ -31,3 +31,6 @@ lerna-debug.log*
 
 # misc
 .DS_Store
+
+# local wrangler config
+wrangler.local.jsonc
