+++ apps/server/README.md
@@ -19,3 +19,5 @@ Pass the `CloudflareBindings` as generics when instantiation `Hono`:
 // src/index.ts
 const app = new Hono<{ Bindings: CloudflareBindings }>()
 ```
+
+For development purpose use `wrangler.local.jsonc`
